requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
pytest
psutil
prometheus-client